Rename konduit-serving-javacv to konduit-serving-image; add height/width to Image API
