+++ konduit-serving-data/konduit-serving-image/pom.xml
