+++ konduit-serving-data/konduit-serving-image/src/test/resources/data/5_32x32.png
