--- konduit-serving-data/konduit-serving-javacv/src/main/resources/META-INF/services/ai.konduit.serving.pipeline.api.format.ImageConverter
@@ -1,30 +1,0 @@
-#
-# /* ******************************************************************************
-#  * Copyright (c) 2020 Konduit K.K.
-#  *
-#  * This program and the accompanying materials are made available under the
-#  * terms of the Apache License, Version 2.0 which is available at
-#  * https://www.apache.org/licenses/LICENSE-2.0.
-#  *
-#  * Unless required by applicable law or agreed to in writing, software
-#  * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
-#  * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
-#  * License for the specific language governing permissions and limitations
-#  * under the License.
-#  *
-#  * SPDX-License-Identifier: Apache-2.0
-#  ******************************************************************************/
-#
-
-ai.konduit.serving.data.javacv.format.JavaCVImageConverters$FrameToMatConverter
-ai.konduit.serving.data.javacv.format.JavaCVImageConverters$MatToFrameConverter
-ai.konduit.serving.data.javacv.format.JavaCVImageConverters$FrameToOpenCVMatConverter
-ai.konduit.serving.data.javacv.format.JavaCVImageConverters$OpenCVMatToFrameConverter
-ai.konduit.serving.data.javacv.format.JavaCVImageConverters$MatToOpenCVMatConverter
-ai.konduit.serving.data.javacv.format.JavaCVImageConverters$OpenCVMatToMatConverter
-ai.konduit.serving.data.javacv.format.JavaCVImageConverters$MatToPng
-ai.konduit.serving.data.javacv.format.JavaCVImageConverters$PngToMat
-ai.konduit.serving.data.javacv.format.JavaCVImageConverters$FrameToPng
-ai.konduit.serving.data.javacv.format.JavaCVImageConverters$PngToFrame
-ai.konduit.serving.data.javacv.format.JavaCVImageConverters$OpenCVMatToPng
-ai.konduit.serving.data.javacv.format.JavaCVImageConverters$PngToOpenCVMat
