+++ konduit-serving-data/pom.xml
@@ -32,7 +32,7 @@
     <packaging>pom</packaging>
     <modules>
         <module>konduit-serving-nd4j</module>
-        <module>konduit-serving-javacv</module>
+        <module>konduit-serving-image</module>
     </modules>
